examples: simplified the int32 publisher and created own app-example options
