+++ examples/int32_publisher/main/CMakeLists.txt
@@ -1,3 +1,2 @@
-idf_component_register(SRCS rclc_int32_publisher.c main.c 
-                    INCLUDE_DIRS ""
-                    )
+idf_component_register(SRCS main.c  
+                    INCLUDE_DIRS "")
